Update buffer mapping

+++ Samples/Program/Program.cl
@@ -1,5 +1,5 @@
 __kernel
-void Copy(__global const uint* src, __global uint* dst)
+void Copy(__global const uchar* src, __global uchar* dst)
 {
     size_t idx = get_global_id(0);
     dst[idx] = src[idx];
